Add wildcardKey type
Being able to type a known structure ahead of time is super useful. Sometimes, however, this isn't possible. What if
your application has an entirely dynamic portion, where you use data from the server to build, day, a form? In this situation
it'd be great to still be able to type. This commit introduces the basic implementation of the 'wildcardKey' type, for use
in defining an object property. At the moment, every object can have one wildcardKey which specifies a specific type (including any()) which the reducer will happily accept, if an unknown property has it. The aim is to introduce regex to this type, and the ability to use multiple instances of the type, in order to specifically craft dynamic reducer structures (e.g. all properties with _date in them will be assigned to the Type.any(), but all others must be Type.string()).